Solução bug ao deletar sócio

+++ html/socio/sistema/processa_deletar_socio.php
@@ -7,20 +7,16 @@
     }else if(is_string($_POST)){
         $_POST = json_decode($_POST, true);
     }
-    var_dump($_REQUEST);
     $deletado =  false;
     $id = $_POST['id_socio'];
     $pessoa = $_POST['pessoa'];
     if($socio = mysqli_fetch_array(mysqli_query($conexao, "SELECT * FROM `socio` WHERE id_socio = $id"))){
-        echo("teste1");
         $id_pessoa = $socio['id_pessoa'];
-        echo($id_pessoa);
-        echo("DELETE FROM `socio` WHERE id_socio=$id");
-        if(mysqli_query($conexao, "DELETE FROM `socio` WHERE id_socio=$id")){
-            echo("teste2");
-            if(mysqli_query($conexao, "DELETE FROM `pessoa` WHERE id_pessoa=$id_pessoa")){
-                echo("teste3");
-                $deletado = true;
+        if(mysqli_query($conexao, "DELETE FROM `log_contribuicao` WHERE id_socio=$id")){
+            if(mysqli_query($conexao, "DELETE FROM `socio` WHERE id_socio=$id")){
+                if(mysqli_query($conexao, "DELETE FROM `pessoa` WHERE id_pessoa=$id_pessoa")){
+                    $deletado = true;
+                }
             }
         }
     }
